+++ lib/formex/builder.ex
@@ -1,28 +1,15 @@
 defmodule Formex.Builder do
   import Ecto.Changeset
   alias Formex.Form
-  @repo Application.get_env(:formex, :repo)
 
   @moduledoc """
-  The form builder.
+  The form builder to be used in a controller. Imported by `Formex.Controller`.
 
   Example:
 
   ```
   form = create_form(App.ArticleType, %Article{})
   render(conn, "new.html", form: form)
-  ```
-
-  ```
-  App.ArticleType
-  |> create_form(%Article{}, article_params)
-  |> insert_form_data
-  |> case do
-    {:ok, _article} ->
-      # ...
-    {:error, form} ->
-      # ...
-  end
   ```
   """
 
@@ -46,30 +33,6 @@ defmodule Formex.Builder do
     |> type.build_form()
     |> create_changeset(type)
     # |> IO.inspect
-  end
-
-  @doc """
-  Invokes `Repo.insert`. In case of `:error`, returns `{:error, form}` (with new `form.changeset`
-  value) instead of `{:error, changeset}` (as Ecto does)
-  """
-  @spec insert_form_data(Form.t) :: {:ok, Ecto.Schema.t} | {:error, Form.t}
-  def insert_form_data(form) do
-    case @repo.insert(form.changeset) do
-      {:ok, schema}       -> {:ok, schema}
-      {:error, changeset} -> {:error, Map.put(form, :changeset, changeset)}
-    end
-  end
-
-  @doc """
-  Invokes `Repo.update`. In case of `:error`, returns `{:error, form}` (with new `form.changeset`
-  value) instead of `{:error, changeset}` (as Ecto does)
-  """
-  @spec update_form_data(Form.t) :: {:ok, Ecto.Schema.t} | {:error, Form.t}
-  def update_form_data(form) do
-    case @repo.update(form.changeset) do
-      {:ok, schema}       -> {:ok, schema}
-      {:error, changeset} -> {:error, Map.put(form, :changeset, changeset)}
-    end
   end
 
   #